+++ mobile/android/android-components/components/feature/session/src/test/java/mozilla/components/feature/session/TrackingProtectionUseCasesTest.kt
@@ -4,11 +4,19 @@
 
 package mozilla.components.feature.session
 
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import mozilla.components.browser.state.action.CustomTabListAction
 import mozilla.components.browser.state.action.EngineAction
+import mozilla.components.browser.state.action.TabListAction
+import mozilla.components.browser.state.selector.findCustomTab
+import mozilla.components.browser.state.selector.findTab
 import mozilla.components.browser.state.state.BrowserState
 import mozilla.components.browser.state.state.ContentState
 import mozilla.components.browser.state.state.EngineState
 import mozilla.components.browser.state.state.TabSessionState
+import mozilla.components.browser.state.state.TrackingProtectionState
+import mozilla.components.browser.state.state.createCustomTab
+import mozilla.components.browser.state.state.createTab
 import mozilla.components.browser.state.store.BrowserStore
 import mozilla.components.concept.engine.Engine
 import mozilla.components.concept.engine.EngineSession
@@ -17,6 +25,7 @@ import mozilla.components.concept.engine.content.blocking.TrackingProtectionExce
 import mozilla.components.concept.engine.content.blocking.TrackingProtectionExceptionStorage
 import mozilla.components.support.test.any
 import mozilla.components.support.test.ext.joinBlocking
+import mozilla.components.support.test.libstate.ext.waitUntilIdle
 import mozilla.components.support.test.mock
 import mozilla.components.support.test.whenever
 import org.junit.Assert.assertFalse
@@ -25,9 +34,11 @@ import org.junit.Assert.assertNull
 import org.junit.Assert.assertTrue
 import org.junit.Before
 import org.junit.Test
+import org.junit.runner.RunWith
 import org.mockito.Mockito.never
 import org.mockito.Mockito.verify
 
+@RunWith(AndroidJUnit4::class)
 class TrackingProtectionUseCasesTest {
     private lateinit var exceptionStore: TrackingProtectionExceptionStorage
     private lateinit var engine: Engine
@@ -146,10 +157,47 @@ class TrackingProtectionUseCasesTest {
 
     @Test
     fun `remove a tracking protection exception`() {
-        val exception: TrackingProtectionException = mock()
+        val tab1 = createTab("https://www.mozilla.org")
+                .copy(trackingProtection = TrackingProtectionState(ignoredOnTrackingProtection = true))
+
+        val tab2 = createTab("https://wiki.mozilla.org/")
+                .copy(trackingProtection = TrackingProtectionState(ignoredOnTrackingProtection = true))
+
+        val tab3 = createTab("https://www.mozilla.org/en-CA/")
+                .copy(trackingProtection = TrackingProtectionState(ignoredOnTrackingProtection = true))
+
+        val customTab = createCustomTab("https://www.mozilla.org/en-CA/")
+                .copy(trackingProtection = TrackingProtectionState(ignoredOnTrackingProtection = true))
+
+        val exception = object : TrackingProtectionException {
+            override val url: String = tab1.content.url
+        }
+
+        store.dispatch(TabListAction.AddTabAction(tab1)).joinBlocking()
+        store.dispatch(TabListAction.AddTabAction(tab2)).joinBlocking()
+        store.dispatch(TabListAction.AddTabAction(tab3)).joinBlocking()
+        store.dispatch(CustomTabListAction.AddCustomTabAction(customTab)).joinBlocking()
+        store.waitUntilIdle()
+
+        assertTrue(store.state.findTab(tab1.id)!!.trackingProtection.ignoredOnTrackingProtection)
+        assertTrue(store.state.findTab(tab2.id)!!.trackingProtection.ignoredOnTrackingProtection)
+        assertTrue(store.state.findTab(tab3.id)!!.trackingProtection.ignoredOnTrackingProtection)
+        assertTrue(store.state.findCustomTab(customTab.id)!!.trackingProtection.ignoredOnTrackingProtection)
 
         useCases.removeException(exception)
+
         verify(exceptionStore).remove(exception)
+
+        store.waitUntilIdle()
+
+        assertFalse(store.state.findTab(tab1.id)!!.trackingProtection.ignoredOnTrackingProtection)
+
+        // Different domain from tab1 MUST not be affected
+        assertTrue(store.state.findTab(tab2.id)!!.trackingProtection.ignoredOnTrackingProtection)
+
+        // Another tabs with the same domain as tab1 MUST be updated
+        assertFalse(store.state.findTab(tab3.id)!!.trackingProtection.ignoredOnTrackingProtection)
+        assertFalse(store.state.findCustomTab(customTab.id)!!.trackingProtection.ignoredOnTrackingProtection)
     }
 
     @Test